+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
         <activity android:name="me.kirimin.mitsumine.ui.activity.ReadLaterActivity" />
         <activity android:name="me.kirimin.mitsumine.ui.activity.search.KeywordSearchActivity" />
         <activity android:name="me.kirimin.mitsumine.ui.activity.search.UserSearchActivity" />
-        <activity android:name="me.kirimin.mitsumine.ui.activity.SettingsActivity" />
+        <activity android:name="me.kirimin.mitsumine.ui.activity.SettingActivity" />
         <activity android:name="me.kirimin.mitsumine.ui.activity.AboutActivity" />
         <activity android:name="me.kirimin.mitsumine.ui.activity.EntryInfoActivity" />
         <activity android:name=".ui.activity.LoginActivity" />
