+++ app/src/main/java/me/kirimin/mitsumine/ui/activity/SettingActivity.kt
@@ -17,8 +17,9 @@ import android.widget.EditText
 import android.widget.Toast
 
 import kotlinx.android.synthetic.activity_settings.*
+import me.kirimin.mitsumine.ui.fragment.SettingFragment
 
-public class SettingsActivity : ActionBarActivity() {
+public class SettingActivity : ActionBarActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -29,7 +30,7 @@ public class SettingsActivity : ActionBarActivity() {
         actionBar.setDisplayHomeAsUpEnabled(true)
         actionBar.setHomeButtonEnabled(true)
 
-        getFragmentManager().beginTransaction().replace(R.id.content, MyPrefsFragment()).commit()
+        getFragmentManager().beginTransaction().replace(R.id.content, SettingFragment()).commit()
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
@@ -37,65 +38,5 @@ public class SettingsActivity : ActionBarActivity() {
             finish()
         }
         return super.onOptionsItemSelected(item)
-    }
-
-    public class MyPrefsFragment : PreferenceFragment() {
-        override fun onCreate(savedInstanceState: Bundle) {
-            super.onCreate(savedInstanceState)
-            addPreferencesFromResource(R.xml.app_preferences)
-            findPreference("about").setOnPreferenceClickListener {
-                startActivity(Intent(getActivity(), javaClass<AboutActivity>()))
-                false
-            }
-
-            findPreference("ngword").setOnPreferenceClickListener {
-                createEditNGWordDialog().show()
-                false
-            }
-
-            val preference = findPreference("logout")
-            preference.setEnabled(AccountDAO.get() != null)
-            preference.setOnPreferenceClickListener { preference ->
-                AccountDAO.delete()
-                Toast.makeText(getActivity(), getString(R.string.settings_logout_toast), Toast.LENGTH_SHORT).show()
-                preference.setEnabled(false)
-                false
-            }
-        }
-
-        private fun createEditNGWordDialog(): AlertDialog {
-            val ngWordList = NGWordDAO.findAll()
-            return AlertDialog.Builder(getActivity())
-                    .setTitle(R.string.settings_ngword)
-                    .setItems(ngWordList.copyToArray(), { dialog, which ->
-                        if (which == ngWordList.size() - 1) {
-                            createAddDialog().show()
-                        } else {
-                            createDeleteDialog(which).show()
-                        }
-                    })
-                    .create()
-        }
-
-        private fun createAddDialog(): AlertDialog {
-            val editText = EditText(getActivity())
-            return AlertDialog.Builder(getActivity())
-                    .setTitle(R.string.settings_ngword_add)
-                    .setPositiveButton(android.R.string.ok, { dialog, which ->
-                        if (editText.getText().length() != 0) NGWordDAO.save(editText.getText().toString())
-                    })
-                    .setView(editText)
-                    .create()
-        }
-
-        private fun createDeleteDialog(index: Int): AlertDialog {
-            return AlertDialog.Builder(getActivity())
-                    .setTitle(R.string.settings_ngword_delete)
-                    .setPositiveButton(android.R.string.ok, { dialog, which ->
-                        NGWordDAO.delete(NGWordDAO.findAll().get(index))
-                    })
-                    .setNegativeButton(android.R.string.cancel, null)
-                    .create()
-        }
     }
 }
